fix linux_aarch64 stdint redefine

+++ 3rd/math/src/__fpclassify.c
@@ -2,7 +2,7 @@
 
 int __fpclassify(double x)
 {
-	union {double f; uint64_t i;} u = {x};
+	union {double f; unsigned long long i;} u = {x};
 	int e = u.i>>52 & 0x7ff;
 	if (!e) return u.i<<1 ? FP_SUBNORMAL : FP_ZERO;
 	if (e==0x7ff) return u.i<<12 ? FP_NAN : FP_INFINITE;
