+++ 3rd/math/src/fabs.c
@@ -2,7 +2,7 @@
 
 double fabs(double x)
 {
-	union {double f; uint64_t i;} u = {x};
+	union {double f; unsigned long long i;} u = {x};
 	u.i &= ULLONG_NSHIFT/2;
 	return u.f;
 }
