+++ 3rd/math/src/trunc.c
@@ -2,9 +2,9 @@
 
 double trunc(double x)
 {
-	union {double f; uint64_t i;} u = {x};
+	union {double f; unsigned long long i;} u = {x};
 	int e = (int)(u.i >> 52 & 0x7ff) - 0x3ff + 12;
-	uint64_t m;
+	unsigned long long m;
 
 	if (e >= 52 + 12)
 		return x;
